corrected betty error task 2

+++ 0x1A-hash_tables/2-key_index.c
@@ -3,14 +3,15 @@
 /**
  * key_index - gives index of array
  * @key: key
- * @value:  size of the array of the hash table
+ * @size:  size of the array of the hash table
  *
- * Return: the index at which the key/value pair should be stored in the array of the hash table
+ * Return: the index at which the key/value pair
+ * should be stored in the array of the hash table
  */
 
 unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
-	int index;
+	unsigned long int index;
 
 	index = hash_djb2(key) % size;
 	return (index);
